+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -12,6 +12,8 @@ class Versions {
     val slf4j = "2.0.13"
     // https://mvnrepository.com/artifact/ch.qos.logback/logback-classic
     val logback = "1.5.6"
+    // https://mvnrepository.com/artifact/com.google.guava/guava
+    val guava = "33.3.0-jre"
     // https://mvnrepository.com/artifact/org.springframework/spring-core
     val spring = "6.1.11"
     // https://mvnrepository.com/artifact/org.springframework.boot/spring-boot
@@ -24,6 +26,7 @@ class Libraries {
     val derbytools = "org.apache.derby:derbytools:${versions.derby}"
     val slf4j = "org.slf4j:slf4j-api:${versions.slf4j}"
     val logback = "ch.qos.logback:logback-classic:${versions.logback}"
+    val guava = "com.google.guava:guava:${versions.guava}"
     val springContext = "org.springframework:spring-context:${versions.spring}"
     val springBootStarter = "org.springframework.boot:spring-boot-starter:${versions.springBoot}"
     val springBootStarterJdbc = "org.springframework.boot:spring-boot-starter-jdbc:${versions.springBoot}"