+++ jkorm/src/test/kotlin/com/tang/jkorm/session/entity/Account.kt
@@ -1,7 +1,9 @@
 package com.tang.jkorm.session.entity
 
+import com.tang.jkorm.annotation.Column
 import com.tang.jkorm.annotation.Id
-import java.time.LocalDateTime
+import java.math.BigDecimal
+import java.sql.Date
 
 /**
  * @author Tang
@@ -12,11 +14,13 @@ class Account (
     var id: Long? = null,
     var username: String = "",
     var password: String = "",
-    var createTime: LocalDateTime? = null,
-    var balance: Double? = null
+    var createTime: Date? = null,
+    @Column("update_time")
+    var updateTime: Date? = null,
+    var balance: BigDecimal? = null
 
 ) {
     override fun toString(): String {
-        return "Account(id=$id, username='$username', password='$password', createTime=$createTime, balance=$balance)"
+        return "Account(id=$id, username='$username', password='$password', createTime=$createTime, updateTime=$updateTime, balance=$balance)"
     }
 }